requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.19
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.4
python-dotenv==1.0.1
pydantic==2.7.3
pydantic-settings==2.3.3
face-recognition==1.3.0
dlib==19.24.4
Pillow==10.3.0
numpy==1.26.4
opencv-python==4.10.0.82
pytest==8.2.2
httpx==0.27.0
pytest-asyncio==0.23.7
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
black==24.4.2
flake8==7.0.0
mypy==1.10.0
isort==5.13.2
pre-commit==3.7.1